# Don't create cookiejar before each iteration if it should be reused