```
$ benchstat old.txt new.txt
name            old time/op    new time/op    delta
HTTPRequests-4     261µs ±10%     256µs ±11%    ~     (p=0.097 n=18+18)

name            old alloc/op   new alloc/op   delta
HTTPRequests-4    27.7kB ± 0%    27.6kB ± 0%  -0.35%  (p=0.000 n=20+20)

name            old allocs/op  new allocs/op  delta
HTTPRequests-4       368 ± 0%       366 ± 0%  -0.54%  (p=0.000 n=20+20)
```